requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.2.1
rsa==4.9.1
Telethon==1.42.0
yarl==1.22.0
